flask/requirements.txt
librosa==0.10.0.post2
numpy==1.23.5
tensorflow==2.12.0
flask==2.2.3
uwsgi==2.0.20
pyyaml==6.0